refactor(engine): extract models::staging
The ANE staging/bundle-repair concern the ticket's audit missed (#950's plan
correction #1): stage_ane_kokoro_voices, stage_fluidaudio_kokoro_assets,
missing_kokoro_assets(_in), the incomplete-.mlmodelc-bundle scan/purge family,
and the ane_bundle_repair_tests/tts_tests modules that exercise them. It fits
neither download.rs (it downloads via parallel_download, but also walks and
repairs the filesystem cache) nor paths.rs (it resolves no new locations).

mod.rs shrinks to mod/pub-use declarations plus install(), which still calls
staging::purge_incomplete_ane_bundles through a `use`. missing_kokoro_assets
and incomplete_ane_bundle_names are re-exported (`pub use staging::{..}`)
because tts::fluid_kokoro calls them as `crate::models::*` — the split must
not move that call site. stage_ane_kokoro_voices/stage_fluidaudio_kokoro_assets
have no caller outside models:: (only download::download_tts uses them), so
download.rs's import retargets from `super::` to `super::staging::` instead
of adding an unused re-export.

Refs #950
